+++ src/main/java/jenkins/plugins/slack/SlackService.java
@@ -1,6 +1,5 @@
 package jenkins.plugins.slack;
 
-
 import net.sf.json.JSONArray;
 
 public interface SlackService {
@@ -9,4 +8,6 @@ public interface SlackService {
     boolean publish(String message, String color);
 
     boolean publish(String message, JSONArray attachments, String color);
+
+    String getResponseString();
 }