+++ src/main/java/jenkins/plugins/slack/workflow/SlackSendStep.java
@@ -26,7 +26,6 @@ import org.jenkinsci.plugins.workflow.steps.AbstractStepDescriptorImpl;
 import org.jenkinsci.plugins.workflow.steps.AbstractStepImpl;
 import org.jenkinsci.plugins.workflow.steps.AbstractSynchronousNonBlockingStepExecution;
 import org.jenkinsci.plugins.workflow.steps.StepContextParameter;
-
 import org.kohsuke.stapler.AncestorInPath;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.kohsuke.stapler.DataBoundSetter;
@@ -35,7 +34,6 @@ import org.kohsuke.stapler.QueryParameter;
 import javax.annotation.Nonnull;
 import javax.inject.Inject;
 
-
 import static com.cloudbees.plugins.credentials.CredentialsProvider.lookupCredentials;
 
 /**
@@ -43,7 +41,7 @@ import static com.cloudbees.plugins.credentials.CredentialsProvider.lookupCreden
  */
 public class SlackSendStep extends AbstractStepImpl {
 
-    private final @Nonnull String message;
+    private String message;
     private String color;
     private String token;
     private String tokenCredentialId;
@@ -112,7 +110,7 @@ public class SlackSendStep extends AbstractStepImpl {
     @DataBoundSetter
     public void setBaseUrl(String baseUrl) {
         this.baseUrl = Util.fixEmpty(baseUrl);
-        if(this.baseUrl != null && !this.baseUrl.isEmpty() && !this.baseUrl.endsWith("/")) {
+        if (this.baseUrl != null && !this.baseUrl.isEmpty() && !this.baseUrl.endsWith("/")) {
             this.baseUrl += "/";
         }
     }
@@ -136,17 +134,21 @@ public class SlackSendStep extends AbstractStepImpl {
     }
 
     @DataBoundSetter
-    public void setAttachments(String attachments){
+    public void setAttachments(String attachments) {
         this.attachments = attachments;
     }
 
-    public String getAttachments(){
+    public String getAttachments() {
         return attachments;
     }
 
-    @DataBoundConstructor
-    public SlackSendStep(@Nonnull String message) {
+    @DataBoundSetter
+    public void setMessage(String message) {
         this.message = message;
+    }
+
+    @DataBoundConstructor
+    public SlackSendStep() {
     }
 
     @Extension
@@ -170,7 +172,7 @@ public class SlackSendStep extends AbstractStepImpl {
 
             Jenkins jenkins = Jenkins.getActiveInstance();
 
-            if(project == null && !jenkins.hasPermission(Jenkins.ADMINISTER) ||
+            if (project == null && !jenkins.hasPermission(Jenkins.ADMINISTER) ||
                     project != null && !project.hasPermission(Item.EXTENDED_READ)) {
                 return new StandardListBoxModel();
             }
@@ -188,11 +190,12 @@ public class SlackSendStep extends AbstractStepImpl {
         //WARN users that they should not use the plain/exposed token, but rather the token credential id
         public FormValidation doCheckToken(@QueryParameter String value) {
             //always show the warning - TODO investigate if there is a better way to handle this
-            return FormValidation.warning("Exposing your Integration Token is a security risk. Please use the Integration Token Credential ID");
+            return FormValidation
+                    .warning("Exposing your Integration Token is a security risk. Please use the Integration Token Credential ID");
         }
     }
 
-    public static class SlackSendStepExecution extends AbstractSynchronousNonBlockingStepExecution<Void> {
+    public static class SlackSendStepExecution extends AbstractSynchronousNonBlockingStepExecution<SlackResponse> {
 
         private static final long serialVersionUID = 1L;
 
@@ -203,7 +206,7 @@ public class SlackSendStep extends AbstractStepImpl {
         transient TaskListener listener;
 
         @Override
-        protected Void run() throws Exception {
+        protected SlackResponse run() throws Exception {
 
             Jenkins jenkins = Jenkins.getActiveInstance();
 
@@ -212,7 +215,8 @@ public class SlackSendStep extends AbstractStepImpl {
 
             String baseUrl = step.baseUrl != null ? step.baseUrl : slackDesc.getBaseUrl();
             String team = step.teamDomain != null ? step.teamDomain : slackDesc.getTeamDomain();
-            String tokenCredentialId = step.tokenCredentialId != null ? step.tokenCredentialId : slackDesc.getTokenCredentialId();
+            String tokenCredentialId = step.tokenCredentialId != null ? step.tokenCredentialId : slackDesc
+                    .getTokenCredentialId();
             String token;
             boolean botUser;
             if (step.token != null) {
@@ -226,27 +230,28 @@ public class SlackSendStep extends AbstractStepImpl {
             String color = step.color != null ? step.color : "";
 
             //placing in console log to simplify testing of retrieving values from global config or from step field; also used for tests
-            listener.getLogger().println(Messages.SlackSendStepConfig(step.baseUrl == null, step.teamDomain == null, step.token == null, step.channel == null, step.color == null));
+            listener.getLogger().println(Messages
+                    .SlackSendStepConfig(step.baseUrl == null, step.teamDomain == null, step.token == null, step.channel == null, step.color == null));
 
             SlackService slackService = getSlackService(baseUrl, team, token, tokenCredentialId, botUser, channel);
             boolean publishSuccess;
-            if(step.attachments != null){
+            if (step.attachments != null) {
                 JsonSlurper jsonSlurper = new JsonSlurper();
-                JSON json = null;
+                JSON json;
                 try {
                     json = jsonSlurper.parseText(step.attachments);
                 } catch (JSONException e) {
                     listener.error(Messages.NotificationFailedWithException(e));
                     return null;
                 }
-                if(!(json instanceof JSONArray)){
-                    listener.error(Messages.NotificationFailedWithException(new IllegalArgumentException("Attachments must be JSONArray")));
+                if (!(json instanceof JSONArray)) {
+                    listener.error(Messages
+                            .NotificationFailedWithException(new IllegalArgumentException("Attachments must be JSONArray")));
                     return null;
                 }
                 JSONArray jsonArray = (JSONArray) json;
-                for (int i = 0; i < jsonArray.size(); i++) {
-                    Object object = jsonArray.get(i);
-                    if(object instanceof JSONObject){
+                for (Object object : jsonArray) {
+                    if (object instanceof JSONObject) {
                         JSONObject jsonNode = ((JSONObject) object);
                         if (!jsonNode.has("fallback")) {
                             jsonNode.put("fallback", step.message);
@@ -254,15 +259,23 @@ public class SlackSendStep extends AbstractStepImpl {
                     }
                 }
                 publishSuccess = slackService.publish(step.message, jsonArray, color);
-            } else {
+            } else if (step.message != null) {
                 publishSuccess = slackService.publish(step.message, color);
+            } else {
+                listener.error(Messages
+                        .NotificationFailedWithException(new IllegalArgumentException("No message or attachments provided")));
+                return null;
             }
-            if (!publishSuccess && step.failOnError) {
+            SlackResponse response = null;
+            if (publishSuccess) {
+                String responseString = slackService.getResponseString();
+                response = new SlackResponse(responseString);
+            } else if (step.failOnError) {
                 throw new AbortException(Messages.NotificationFailed());
-            } else if (!publishSuccess) {
+            } else {
                 listener.error(Messages.NotificationFailed());
             }
-            return null;
+            return response;
         }
 
         //streamline unit testing